flask-webapp/requirements.txt
alembic==1.4.3
cachelib==0.1.1
click==7.1.2
cssmin==0.2.0
ddtrace==0.45.0
dnspython==2.0.0
email-validator==1.1.2
Faker==5.0.2
Flask==1.1.2
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
intervaltree==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lesscpy==0.14.0
Mako==1.1.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
ply==3.11
protobuf==3.14.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
redis==3.5.3
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.3.22
tenacity==6.3.1
text-unidecode==1.3
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
